backend/agent/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pytest==8.3.4
httpx==0.28.1

# ── Phase C: LangGraph Agent ───────────────────────────────
langgraph==0.4.1
langchain-core==0.3.51
langchain-openai==0.3.12
gitpython==3.1.44
asyncpg==0.30.0
redis[hiredis]==5.2.1
aiofiles==24.1.0

# ── Phase F: Report PDF generation ─────────────────────────
reportlab==4.1.0
